Refactor #42 Product 프로젝트 구조 변경
- ProductAttribution 을 Product안으로 이동
- db 패키지에 entity, repository 묶음
- dto, entity 변환 및 비지니스 로직을 담당하는 business에 converter, dto 패키지 이동
