+++ src/main/java/com/careerbuilder/careerbuilder/domain/product/db/entity/ProductAttribution.java
@@ -1,7 +1,6 @@
-package com.careerbuilder.careerbuilder.domain.productattribution.entity;
+package com.careerbuilder.careerbuilder.domain.product.db.entity;
 
 import com.careerbuilder.careerbuilder.domain.attribution.entity.Attribution;
-import com.careerbuilder.careerbuilder.domain.product.entity.Product;
 import com.careerbuilder.careerbuilder.global.common.baseentity.BaseEntity;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
