+++ src/main/java/com/careerbuilder/careerbuilder/domain/product/db/repository/ProductAttributionRepository.java
@@ -1,6 +1,6 @@
-package com.careerbuilder.careerbuilder.domain.productattribution.repository;
+package com.careerbuilder.careerbuilder.domain.product.db.repository;
 
-import com.careerbuilder.careerbuilder.domain.productattribution.entity.ProductAttribution;
+import com.careerbuilder.careerbuilder.domain.product.db.entity.ProductAttribution;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
