+++ src/main/java/com/careerbuilder/careerbuilder/domain/product/db/repository/ProductRepository.java
@@ -1,0 +1,15 @@
+package com.careerbuilder.careerbuilder.domain.product.db.repository;
+
+import com.careerbuilder.careerbuilder.domain.product.db.entity.Product;
+import com.careerbuilder.careerbuilder.domain.product.db.repository.querydsl.ProductRepositoryCustom;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface ProductRepository extends JpaRepository<Product, Long>, ProductRepositoryCustom {
+
+    Optional<Product> searchByNameContainingIgnoreCase(String keyword);
+
+}
