+++ src/main/java/com/careerbuilder/careerbuilder/domain/product/db/repository/querydsl/ProductRepositoryCustom.java
@@ -1,0 +1,6 @@
+package com.careerbuilder.careerbuilder.domain.product.db.repository.querydsl;
+
+public interface ProductRepositoryCustom {
+
+}
+
