+++ src/main/java/com/careerbuilder/careerbuilder/domain/product/db/repository/querydsl/ProductRepositoryImpl.java
@@ -1,0 +1,12 @@
+package com.careerbuilder.careerbuilder.domain.product.db.repository.querydsl;
+
+import com.querydsl.jpa.impl.JPAQueryFactory;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
+
+@Repository
+@RequiredArgsConstructor
+public class ProductRepositoryImpl implements ProductRepositoryCustom {
+    private final JPAQueryFactory jpaQueryFactory;
+
+}
