+++ src/main/java/com/careerbuilder/careerbuilder/domain/stock/dto/StockDetailResponse.java
@@ -1,14 +1,12 @@
 package com.careerbuilder.careerbuilder.domain.stock.dto;
 
 import com.careerbuilder.careerbuilder.domain.location.dto.LocationResponse;
-import com.careerbuilder.careerbuilder.domain.product.dto.ProductResponse;
-import com.careerbuilder.careerbuilder.domain.product.entity.Product;
+import com.careerbuilder.careerbuilder.domain.product.business.dto.ProductResponse;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.List;
 import java.util.Map;
 
 @Data
