+++ src/main/java/com/careerbuilder/careerbuilder/domain/stock/entity/Stock.java
@@ -1,7 +1,5 @@
 package com.careerbuilder.careerbuilder.domain.stock.entity;
 
-import com.careerbuilder.careerbuilder.domain.location.entity.Location;
-import com.careerbuilder.careerbuilder.domain.product.entity.Product;
 import com.careerbuilder.careerbuilder.global.common.baseentity.BaseEntity;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
