+++ src/main/java/com/careerbuilder/careerbuilder/domain/transaction/dto/TransactionDetailWithProductListResponse.java
@@ -2,7 +2,7 @@ package com.careerbuilder.careerbuilder.domain.transaction.dto;
 
 import com.careerbuilder.careerbuilder.domain.location.dto.LocationResponse;
 import com.careerbuilder.careerbuilder.domain.partner.dto.PartnerResponse;
-import com.careerbuilder.careerbuilder.domain.product.dto.ProductResponse;
+import com.careerbuilder.careerbuilder.domain.product.business.dto.ProductResponse;
 import com.careerbuilder.careerbuilder.domain.transaction.entity.type.TransactionStatus;
 import com.careerbuilder.careerbuilder.domain.transaction.entity.type.TransactionType;
 import lombok.AllArgsConstructor;
